Add more details about specifying a CA pin
Closes #9946

The CA pin is the only Teleport configuration field where we have not
yet documented the possibility of specifying a value using a file path.
This change includes this information, as well as the (also
undocumented) fact that you can specify a list of CA pins. Also updates
the docs path that the CA pin reference links to.