fix: correct minSupportedLabelExpressionVersion to v13.1.1 (#27894)
Label expressions went out in v13.1.1, this commit corrects the
`minSupportedLabelExpressionVersion` constant to match.
This is necessary for label expressions to work when nodes on versions
between 13.1.1 and 13.2.0 connect to auth servers on versions
>=14.0.0-alpha.
This doesn't affect any released versions, we just want this to be right
before we ship v14.

This was expected, so testcases already use
`minSupportedLabelExpressionVersion` and are automatically up to date.