[v13] Fixes the "Run as different user" window freezing (#27874)
* Fixes the "Run as different user" window freezing

The problem was that we were occasionally getting SCARD_IOCTL_GETSTATUSCHANGEW
calls with infinite timeouts, which we would never return from because our
status never changes. This is the correct behavior in the short run, but we
were ignoring the meaning of SCARD_IOCTL_CANCEL, which tells such long-hanging
calls to return with an SCARD_E_CANCELLED return value.

This was causing the "Run as different user" window to freeze, because something
internally was waiting around for an SCARD_E_CANCELLED that would never come.

To fix this, now whenever we get an SCARD_IOCTL_GETSTATUSCHANGEW with an infinite
timeout, we add to the corresponding context a DeviceControlResponse that wraps
the necessary SCARD_E_CANCELLED return value. Later, when an SCARD_IOCTL_CANCEL
is received for that context, we grab the DeviceControlResponse and return it.

* removes unecessary PartialEq

* Removes unnecessary Option

* consolidate and improve Context.get error handling/message

* change [#allow(dead_code)] to #[cfg(test)]

* updates comment