Make Teleport config instructions easier to follow (#27968)
In the `docs/pages/includes/tls-certificate-setup.mdx` partial, one code
snippet includes an escape character after a pipe in order to render a
shell command across two lines. For users who copy the command manually,
this can lead to unexpected results. This change removes the escape
character and renders the command on one line.

This change also makes the use of `Var` components in code snippets more
explicit so users aren't tempted to paste the example command into, say,
a browser address bar in order to change the placeholder values. (This
is what led to issues with the escape character to begin with.)