ci: Add dummy workflow for linux builds (#1696)
Add a dummy workflow on master for migrating linux builds from drone to
GitHub Actions. Dispatchable workflows must exist on master to be
called, after which a branch/PR can modify that workflow for testing
during development.

This is separate to the `release-linux-arm64.yml` workflow as it is for
direct migration from drone, and not migration to centos7 builders. That
will come later.

Source-commit: gravitational/teleport.e@310ea5d17c4b06f0d4b45d515b4d4888c2a60c3e
Source-pull-requests: gravitational/teleport.e#1696