Remove further unused fips infrastructure (#27900)
* Remove Dockerfile-arm-fips

We don't build fips for arm, as documented in
https://github.com/gravitational/teleport/issues/10581.

* Stop building buildbox-fips

We do not use this buildbox for anything. This step is failing because
the supporting infrastructure for buildbox-fips was removed in
https://github.com/gravitational/teleport/pull/26859.

* Fix fips buildbox

BUILDBOX_FIPS was removed, replaced by BUILDBOX_CENTOS7_FIPS.
Unfortunately I missed updating this target in #26859.
This subsequently broke e CI.