[v12] Docs: Update Okta SSO Guide (#27951)
* make dynamic edit first

* expand scope...

So that dynamic resource is still the default option for all scopes

* add local auth tip...

Since we're telling the reader to switch to SAML before it's configured, we should provide a way to log in again using local auth.

* provide full path to key...

Since updated docs using <code>tctl sso configure</code> won't always include the key for easy reference in its output.

* update okta guide to use tctl sso configure

* Specify enterprise/cloud tctl downloads...

Because <code>tctl sso configure saml</code> is only available in the enterprise tctl versions. This makes this PR dependent on #26124

* strongly encourage testing

* use preset flag