Make Plugin List show Authentication error (#1715)
Make Plugin List show Authentication error

Then a plugin exits with error, any plugin status in the UI is.
Immediately overwritten by an "Unknown error" status, which is unhelpful
for troubleshooting.

This patch preserves the plugin state at exit, if and only if is in
AccessDenied state) on exit. This at least gives the user a hint that their
access setting might be wrong.

Source-commit: gravitational/teleport.e@7af29f7ae0cc9a6babc7877ff6895712ae87919e
Source-pull-requests: gravitational/teleport.e#1715