Adds automatic approver user for use with access plugins (#27829)
* Adds automatic approver user for use with access plugins

Adds an internal user `@teleport-access-approval-bot` with the associated
role `@teleport-access-approver`. This new role has grants the right to
approve any and all role access requests, and both the user and role
resources are automatically created on startup if not already present.

While the `@teleport-access-approval-bot` should under no circumstances
be allowed to log into the cluster or appear in UI user listings of any
kind (except as the approver in an access request), this PR does not
implement any or the requisite filtering or enforcement. This is coming
in a future PR.

* Don't crash when building OSS

* fix test fixup

* Adds labels to the Preset approval bot user

Also ads basic update functionality for preset users, and tests
for same.

* typo fix

* Commentary

* linter appeasement

* Update tests

* Update docs

* Post-merge cleanup

* Update lib/auth/init_test.go

Co-authored-by: Marco André Dinis <marco.dinis@goteleport.com>

* Address review feedback

* Filter out user in WebUI and plumb Teleport user through

---------

Co-authored-by: Marco André Dinis <marco.dinis@goteleport.com>