Respect client idle timeout setting (#27484) (#28227)
Client connections were never using the `srv.TrackingReadConn`
created by the `srv.ConnectionMonitor` which caused all connections to
appear idle and be disconnected whenever the client idle timeout
was reached.

`(srv.ConnectionMonitor) MonitorConn` now returns the provided connection
which is wrapped in the tracking connection. Callers must ensure that
the returned connection is used instead of the connection passed in
to `MonitorConn`.

Fixes #27392