Reduce debug log spam for TeleportReady events (#28320)
In #28229 we added some additional logging to help troubleshoot
when Teleport is unable to become ready. This can be a little
bit too noisy if a completely unrelated event is broadcast.

In this change, we return an error only when we have something
useful to say, and rely on a separate boolean return value to
indicate whether or not the mapping is satisfied.