Teleport One Off Script (#27852)
* Teleport One Off Script

This PR creates an endpoint that returns a script.
This script does the following:
- detects the correct os/arch, downloads and extracts the teleport
  binary package
- runs teleport with arguments defined in the endpoint's query param

This will allow the Discover wizard to give bash scripts in the form of
`bash -c "$(curl ..../oneoff... )"` to the user.
The user is then asked to run those commands.

One use case for those commands is the creation of specific AWS Roles
which are then used to give Teleport extra access to other AWS Services
/ Actions.
This specific command will come in a later PR.

* remove endpoint

* disable shellcheck for go templated vars

* use bintest v3