Make exit_on_sync work as advertised (#1740)
The `exit_on_sync` toggle doesn't work as advertised as many other (then
superfluous) services will keep Teleport running. This takes the "dumb" approach
and asks Teleport to shutdown when the flag is set. Note that the flag is
exclusive to "fileconf", so the behavior can't be triggered via Cloud resources.
It is assumed that configs with this flag are standalone Jamf service
configurations.

RFD reference:
https://github.com/gravitational/teleport.e/blob/master/rfd/0007e-device-trust-mdm-integration.md#jamf-inventory-sync.

#826

Source-commit: gravitational/teleport.e@792fd26f9c478fd78c7cf6f97eb44e9ae1a7a806
Source-pull-requests: gravitational/teleport.e#1740