Add security notes to the session recording guide (#28230)
* Add security notes to the session recording guide

While there are several security risks for using session recording for
SSH, these are only documented in the guide for Enhanced Session
Recording, making them easy to miss if you are only reading the session
recording guide in the "Architecture" section of the docs.

The session recording architecture guide and the Enhanced Session
Recording guide get roughly equivalent traffic, so this change adds a
variation on the security warning in the Enhanced Session Recording
guide to the session recording architecture guide, ensuring that the
security guidance is visible in both locations.

* Respond to wadells feedback