saml.go: Prevent Large Compressed Responses
It is possible to efficiently compress large repeating values.  This could result in the service exhausting its memory in a denial of service as described in this issue: https://github.com/gravitational/teleport-private/issues/659

This change limits the compressed SAML responses to 1MB, at which size an error is immediately returned.

In addition to added test cases, seeds were added to `fuzz_test.go`.

Source-commit: gravitational/teleport.e@21cb950bcb4d1879152e10b4b01163279ab3fc4e