Ensure that all scoped components match forScopes
For nearly all scoped components in the docs, the value of the `scope`
attribute matches the scopes configured in the `forScopes` field for the
component's page.

This change corrects the components that _don't_ match a page's `forScopes`
configuration. This anticipates a new docs linter to check the scopes of
each component.