Edit playbook user in the Ansible guide (#28790)
Closes #16612

The guide names the Machine ID getting started guide as a prerequisite,
and the getting started guide shows how to authorize Machine ID to log
in as `root`.

This change edits the playbook example in the Ansible guide to use
`root`, and uses a `Var` component in case the user configured Machine
ID to have another login instead.