Add the AccessList to the cache.
The AccessList has been added into the cache. In order to support this, the
concept of wrappers has been added to api/client/events.go. Since events
consist of gogo messages and the new AccessList message is not using gogo,
a wrapper around the message has been created such that the message is
created in both gogo and regular protobuf.

This involves creating a new set of packages in api/gen/wrappers and then
modifying buf-gogo.gen.yaml to replace references to these wrapper packages
with the aforementioned packages. This should hopefully be temporary while
we attempt to excise gogo from the codebase.

Protoadapt is not usable for munging between regular protobuf and gogo
protobuf as gogo does not support V2 protobuf messages, so instead the
messages are marshaled and unmarshaled into and from the raw wire format,
which works as expected.