Events and supporting objects to non-Gogo.
In order to support the new access list being added to the cache, the
events object and its supporting objects needs to be non-Gogo, as the
access list object does not have the Gogo functions required by the
Gogo version of the `Event` object.

To support this, the `Operation`, `Event`, and `Watch` messages have been
migrated into `api/proto/teleport/event/v1` along with the `WatchEvents`
and `WatchPendingHeadlessAuthentications` RPC calls, both of which use
these objects. The API client `AuthServiceClient` has been given the
new `EvenServiceClient` to maintain compatibility, and the various
related functions have been updated to properly register this new
gRPC service.

To test this, a version of tsh with and without these changes was used
to create and wait for access requests, and plugins were registered and
deleted. Both of these use a stream watcher, which relies on the services
that were migrated. This worked properly for both version of tsh.
Additionally, this was tested against a v13 tsh, which also worked as
expected.

`WatchPendingHeadlessAuthentications` is not currently testable because
it doesn't appear to be actually used by anything.