ci: Switch MacOS build to ARM64 runners (#1790)
The ARM64 runners seem to be about 50% faster. Faster is better and
cheaper. Switch to these.

NOTE: At the moment these runners are in beta and currently have a bug
where `~/.cache` and `~/.config` are root-owned. We have a step to fix
this that will be removed when the runners are fixed. Also `pkg-config`
is missing so we install that with brew.

Source-commit: gravitational/teleport.e@456473f6df7f648486a6ad1abc2601aa440c7ed4
Source-pull-requests: gravitational/teleport.e#1790