dronegen: Switch linux-based push builds to GitHub (#28805)
* dronegen: Switch linux-based push builds to GitHub

Change the drone pipelines for linux-based push builds to call a GitHub
actions workflow instead of running on drone runners. This includes one
of the builds for Windows which is done in a Linux container.

The old push pipelines that run the build on drone runners is now
removed as it is no longer used.

* ci: Update .drone.yml

This updates the linux-based push pipelines to delegate to github
actions.

* Update e ref for push linux workflows