tctl alerts ack: Make --reason optional (#28953)
This fixes `tctl alerts ack ls` which used to not work due to the reasons
described in the comment.

Providing a reason is still required. The only difference is that instead
of having the CLI fail immediately if the flag is missing, the CLI will
issue a request to the cluster which will fail due to a missing reason.