Add a "5 trusted devices" limit for usage-based accounts (#1801)
Add a limit for usage-based accounts, such as Teleport Team accounts.

The device trust feature, introduced by
https://github.com/gravitational/teleport/pull/28919, is now fully checked here.
This includes the top-level toggle and devices limit.

I've opted to count devices on enrollment rather than registration. This is the
more permissive option, as it lets admins register their inventory in
preparation for an account upgrade. Full-scanning devices isn't the best, but we
don't have many options and the limits/number of devices for such accounts are
expected to be low.

https://github.com/gravitational/cloud/issues/5148

Source-commit: gravitational/teleport.e@c420cfcb955b2c87919416ce2989a067c0603218
Source-pull-requests: gravitational/teleport.e#1801