[v13] fix: suppress search events
Backport #29032 to branch/v13

This PR suppresses the emission of Resource Access Request Search events
when the search does not actually make use of any additional roles the
user doesn't already have, most commonly when they have no
search_as_roles or preview_as_roles configured at all.

The reasoning is that these searches often happen automatically, when
the user merely navigates to the Applications page or the Access
Requests page, and there is no reason to spam the audit events when
there are no extra privileges being assumed for the search.

In light of the fact these events are often emitted when no access
request is actually being created, the displayed name of the events is
also changed to Resource Access Search here.

I also took this opportunity to be a bit more defensive about
auth.(*Context).UseExtraRoles, which previously would elevate the auth
context's roles for the duration of the RPC. I could see this leading to
a bug if another call is ever chained after ListResources within a
single RPC. I opted to replace this with auth.(*Context).WithExtraRoles
which returns a new authz.(*Context) with the additional roles added.
ListResources now defers a call to restore the original auth context
before it returns.