[v13] Add more details about specifying a CA pin (#28886)
* Add more details about specifying a CA pin

Closes #9946

The CA pin is the only Teleport configuration field where we have not
yet documented the possibility of specifying a value using a file path.
This change includes this information, as well as the (also
undocumented) fact that you can specify a list of CA pins. Also updates
the docs path that the CA pin reference links to.

* Respond to zmb3 feedback

* Respond to lsgunn-teleport feedback