Updates buildbox `osslsigncode` to 2.6 (#29043)
OpenSSL disabled some legacy algorithms when moving to 3.0. The tool
we use for Windows Code Signing - `osslsigncode` - recently upgraded to
OpenSSL >= 3.0, which broke our Windows Code Sigining path as the
Windows Code Signing certificate certificate depends on at least
one of these legacy algorithms.

The osslsigncode progect has published a release that fixed this issue,
but that release has not made it through to Ubuntu yet; hence we are
manyally downloading and installing the tool from github during the
buildbox image creation.

See-Also: #28722