Add support for PAM handling in terminal
This commit adds the ability for the process to handle PAM (Pluggable Authentication Modules) within the terminal. Specifically, a new file "ContinueWriteFile" was added to handle the PAM context and ensure continued execution of the process post PAM pre-processing. In addition, a 'WaitForPam' method was added in terminal handling logic to manage and wait for PAM tasks. This was important to ensure controlled execution and improved handling of session management. Further, this also includes initial plumbing to ensure that sessions with PAM are launched correctly within enhanced recording contexts.