v11: dronegen: Switch linux-based push builds to GitHub (#29307)
Change the drone pipelines for linux-based push builds to call a GitHub
actions workflow instead of running on drone runners. This includes one
of the builds for Windows which is done in a Linux container.

The old push pipelines that run the build on drone runners is now
removed as it is no longer used.

Update `.drone.yml` to use the new GitHub actions push pipelines.

Update e ref for push linux workflows.