Add in user login state. (#29365)
The user login state has been introduced, which will house ephemeral data
associated with users that lives separately from the fixed user definition.
This object will only live on the auth server and only be used during very
specific circumstances, so for now it will not be cached and it will not be
exposed via the API. As such, the object and its backend have been introduced
in one commit.

---------

Co-authored-by: Alan Parra <alan.parra@goteleport.com>