Remove dependency of etcd from tctl (#29377) (#29394)
The following import tree resulted in tctl importing etcd:

| github.com/gravitational/teleport/tool/tctl
|--> github.com/gravitational/teleport/tool/tctl/common
|---> github.com/gravitational/teleport/lib/auth
|----> github.com/gravitational/teleport/lib/auth/okta
|-----> github.com/gravitational/teleport/lib/services/local
|-------> github.com/gravitational/teleport/lib/backend/etcdbk
|--------> go.etcd.io/etcd/client/v3

This stems from the assistant service which checks if the backend
is etcd by using type assertion. A new GetName method is added
to backend.Backend so that the storage mechanism can be determined
via string comparison instead of importing backend implementations
to do type assertion.